12-09-23

+++ average.java
@@ -1,7 +1,9 @@
+//average of elements in an array
+
 class average{
 	public static void main(String args[]){
 		int array[] = {2,3,5,7};
-		int result = 0;
+		float result = 0;
 		int i;
 		for(i=0;i<4;i++)
 			result = result + array[i];
